refactor(backend): complete Tooling domain ownership
